src/upstage/test/test_geography/test_conversions.py
# Copyright (C) 2024 by the Georgia Tech Research Institute (GTRI)

# Licensed under the BSD 3-Clause License.
# See the LICENSE file in the project root for complete license terms and disclaimers.

import pytest

from upstage.geography import conversions, spherical, wgs84

SC = conversions.SphericalConversions
WSGC = conversions.WGS84Conversions
SC2 = spherical.Spherical
WSGC2 = wgs84.WGS84


@pytest.mark.parametrize("use", [SC, SC2, WSGC, WSGC2])
def test_conversions(use, random_lla):
    # Do a back and forth test of random Lat Lon Alt
    ecef = use.lla2ecef(random_lla)
    lla_from_ecef = use.ecef2lla(ecef)
    for a, b in zip(lla_from_ecef, random_lla):
        assert pytest.approx(a) == b
